backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Environment & Config
python-dotenv==1.0.0

# CORS & Security
python-multipart==0.0.6

# AWS S3 (for Day 12-13)
boto3==1.34.0

# Web Scraping (for Day 12-13)
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# LLM APIs (for Day 14)
openai==1.3.7
anthropic==0.7.0

# Notifications (for Day 15-16)
twilio==8.11.0

# Scheduling (for Day 17)
apscheduler==3.10.4

# Utilities
python-dateutil==2.8.2